Moves around files in com.meti.util
