+++ docs/javadocs/com/meti/server/SocketHandler.html
@@ -139,7 +139,7 @@ implements <a href="../../../com/meti/util/Handler.html" title="interface in com
                             </tr>
                             <tr class="altColor">
                                 <td class="colOne"><code><span class="memberNameLink"><a
-                                        href="../../../com/meti/server/SocketHandler.html#SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.Callback-java.net.Socket-com.meti.server.Server-">SocketHandler</a></span>(java.util.concurrent.ExecutorService&nbsp;executorService,
+                                        href="../../../com/meti/server/SocketHandler.html#SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.intf.Callback-java.net.Socket-com.meti.server.Server-">SocketHandler</a></span>(java.util.concurrent.ExecutorService&nbsp;executorService,
                                     <a href="../../../com/meti/util/Callback.html" title="interface in com.meti.util">Callback</a>&lt;java.lang.Exception&gt;&nbsp;exceptionCallback,
                                     java.net.Socket&nbsp;socket,
                                     <a href="../../../com/meti/server/Server.html" title="class in com.meti.server">Server</a>&nbsp;server)</code>&nbsp;
@@ -195,7 +195,7 @@ implements <a href="../../../com/meti/util/Handler.html" title="interface in com
                         <!--   -->
                     </a>
                         <h3>Constructor Detail</h3>
-                        <a name="SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.Callback-java.net.Socket-com.meti.server.Server-">
+                        <a name="SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.intf.Callback-java.net.Socket-com.meti.server.Server-">
                             <!--   -->
                         </a>
                         <ul class="blockListLast">
