+++ docs/javadocs/com/meti/util/Cargo.html
@@ -103,7 +103,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.util.Cargo</li>
+                <li>com.meti.util.io.Cargo</li>
             </ul>
         </li>
     </ul>
@@ -123,7 +123,7 @@ implements java.io.Serializable</pre>
                     <dt><span class="simpleTagLabel">Since:</span></dt>
                     <dd>12/3/2017</dd>
                     <dt><span class="seeLabel">See Also:</span></dt>
-                    <dd><a href="../../../serialized-form.html#com.meti.util.Cargo">Serialized Form</a></dd>
+                    <dd><a href="../../../serialized-form.html#com.meti.util.io.Cargo">Serialized Form</a></dd>
                 </dl>
             </li>
         </ul>
