+++ docs/javadocs/com/meti/util/Change.html
@@ -97,7 +97,7 @@
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.util.Change</li>
+                <li>com.meti.util.io.Change</li>
             </ul>
         </li>
     </ul>
