+++ docs/javadocs/com/meti/util/Dialog.html
@@ -103,7 +103,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.util.Dialog</li>
+                <li>com.meti.util.fx.Dialog</li>
             </ul>
         </li>
     </ul>
