+++ docs/javadocs/com/meti/util/LoggerHandler.html
@@ -106,7 +106,7 @@ var activeTableTab = "activeTableTab";
                 <li>java.util.logging.Handler</li>
                 <li>
                     <ul class="inheritance">
-                        <li>com.meti.util.LoggerHandler</li>
+                        <li>com.meti.util.misc.LoggerHandler</li>
                     </ul>
                 </li>
             </ul>
