+++ docs/javadocs/com/meti/util/Loop.html
@@ -103,7 +103,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.util.Loop</li>
+                <li>com.meti.util.concurrent.Loop</li>
             </ul>
         </li>
     </ul>
@@ -176,7 +176,7 @@ implements java.lang.Runnable</pre>
                             </tr>
                             <tr class="rowColor">
                                 <td class="colOne"><code><span class="memberNameLink"><a
-                                        href="../../../com/meti/util/Loop.html#Loop-com.meti.util.Callback-">Loop</a></span>(<a
+                                        href="../../../com/meti/util/Loop.html#Loop-com.meti.util.intf.Callback-">Loop</a></span>(<a
                                         href="../../../com/meti/util/Callback.html" title="interface in com.meti.util">Callback</a>&lt;java.lang.Exception&gt;&nbsp;exceptionCallback)</code>&nbsp;
                                 </td>
                             </tr>
@@ -271,7 +271,7 @@ implements java.lang.Runnable</pre>
                                 <pre>public&nbsp;Loop()</pre>
                             </li>
                         </ul>
-                        <a name="Loop-com.meti.util.Callback-">
+                        <a name="Loop-com.meti.util.intf.Callback-">
                             <!--   -->
                         </a>
                         <ul class="blockListLast">
