+++ docs/javadocs/com/meti/util/Queuer.html
@@ -78,7 +78,7 @@ var activeTableTab = "activeTableTab";
         <ul class="subNavList">
             <li>Summary:&nbsp;</li>
             <li>Nested&nbsp;|&nbsp;</li>
-            <li><a href="#fields.inherited.from.class.com.meti.util.Loop">Field</a>&nbsp;|&nbsp;</li>
+            <li><a href="#fields.inherited.from.class.com.meti.util.concurrent.Loop">Field</a>&nbsp;|&nbsp;</li>
             <li><a href="#constructor.summary">Constr</a>&nbsp;|&nbsp;</li>
             <li><a href="#method.summary">Method</a></li>
         </ul>
@@ -103,11 +103,11 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li><a href="../../../com/meti/util/Loop.html" title="class in com.meti.util">com.meti.util.Loop</a>
+                <li><a href="../../../com/meti/util/Loop.html" title="class in com.meti.util">com.meti.util.concurrent.Loop</a>
                 </li>
                 <li>
                     <ul class="inheritance">
-                        <li>com.meti.util.Queuer</li>
+                        <li>com.meti.util.concurrent.Queuer</li>
                     </ul>
                 </li>
             </ul>
@@ -141,7 +141,7 @@ extends <a href="../../../com/meti/util/Loop.html" title="class in com.meti.util
                     </a>
                         <h3>Field Summary</h3>
                         <ul class="blockList">
-                            <li class="blockList"><a name="fields.inherited.from.class.com.meti.util.Loop">
+                            <li class="blockList"><a name="fields.inherited.from.class.com.meti.util.concurrent.Loop">
                                 <!--   -->
                             </a>
                                 <h3>Fields inherited from class&nbsp;com.meti.util.<a
@@ -211,7 +211,7 @@ extends <a href="../../../com/meti/util/Loop.html" title="class in com.meti.util
                             </tr>
                         </table>
                         <ul class="blockList">
-                            <li class="blockList"><a name="methods.inherited.from.class.com.meti.util.Loop">
+                            <li class="blockList"><a name="methods.inherited.from.class.com.meti.util.concurrent.Loop">
                                 <!--   -->
                             </a>
                                 <h3>Methods inherited from class&nbsp;com.meti.util.<a
@@ -353,7 +353,7 @@ extends <a href="../../../com/meti/util/Loop.html" title="class in com.meti.util
         <ul class="subNavList">
             <li>Summary:&nbsp;</li>
             <li>Nested&nbsp;|&nbsp;</li>
-            <li><a href="#fields.inherited.from.class.com.meti.util.Loop">Field</a>&nbsp;|&nbsp;</li>
+            <li><a href="#fields.inherited.from.class.com.meti.util.concurrent.Loop">Field</a>&nbsp;|&nbsp;</li>
             <li><a href="#constructor.summary">Constr</a>&nbsp;|&nbsp;</li>
             <li><a href="#method.summary">Method</a></li>
         </ul>
