+++ docs/javadocs/com/meti/util/Utility.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.util.Utility</li>
+                <li>com.meti.util.misc.Utility</li>
             </ul>
         </li>
     </ul>
