+++ docs/javadocs/com/meti/util/package-tree.html
@@ -116,7 +116,7 @@
     </ul>
     <h2 title="Interface Hierarchy">Interface Hierarchy</h2>
     <ul>
-        <li type="circle">com.meti.util.Action&lt;T&gt;
+        <li type="circle">com.meti.util.intf.Action&lt;T&gt;
             <ul>
                 <li type="circle">com.meti.util.<a href="../../../com/meti/util/Callback.html"
                                                    title="interface in com.meti.util"><span class="typeNameLink">Callback</span></a>&lt;T&gt;
