+++ docs/javadocs/index-files/index-10.html
@@ -131,7 +131,7 @@
         <dt><span class="memberNameLink"><a href="../com/meti/util/Loop.html#Loop--">Loop()</a></span> - Constructor for
             class com.meti.util.<a href="../com/meti/util/Loop.html" title="class in com.meti.util">Loop</a></dt>
         <dd>&nbsp;</dd>
-        <dt><span class="memberNameLink"><a href="../com/meti/util/Loop.html#Loop-com.meti.util.Callback-">Loop(Callback&lt;Exception&gt;)</a></span>
+        <dt><span class="memberNameLink"><a href="../com/meti/util/Loop.html#Loop-com.meti.util.intf.Callback-">Loop(Callback&lt;Exception&gt;)</a></span>
             - Constructor for class com.meti.util.<a href="../com/meti/util/Loop.html" title="class in com.meti.util">Loop</a>
         </dt>
         <dd>&nbsp;</dd>
