+++ docs/javadocs/index-files/index-16.html
@@ -161,7 +161,7 @@
             - Class in <a href="../com/meti/server/package-summary.html">com.meti.server</a></dt>
         <dd>&nbsp;</dd>
         <dt><span class="memberNameLink"><a
-                href="../com/meti/server/SocketHandler.html#SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.Callback-java.net.Socket-com.meti.server.Server-">SocketHandler(ExecutorService, Callback&lt;Exception&gt;, Socket, Server)</a></span>
+                href="../com/meti/server/SocketHandler.html#SocketHandler-java.util.concurrent.ExecutorService-com.meti.util.intf.Callback-java.net.Socket-com.meti.server.Server-">SocketHandler(ExecutorService, Callback&lt;Exception&gt;, Socket, Server)</a></span>
             - Constructor for class com.meti.server.<a href="../com/meti/server/SocketHandler.html"
                                                        title="class in com.meti.server">SocketHandler</a></dt>
         <dd>&nbsp;</dd>
