+++ docs/javadocs/serialized-form.html
@@ -130,10 +130,11 @@
         <li class="blockList">
             <h2 title="Package">Package&nbsp;com.meti.util</h2>
             <ul class="blockList">
-                <li class="blockList"><a name="com.meti.util.Cargo">
+                <li class="blockList"><a name="com.meti.util.io.Cargo">
                     <!--   -->
                 </a>
-                    <h3>Class <a href="com/meti/util/Cargo.html" title="class in com.meti.util">com.meti.util.Cargo</a>
+                    <h3>Class <a href="com/meti/util/Cargo.html"
+                                 title="class in com.meti.util">com.meti.util.io.Cargo</a>
                         extends java.lang.Object implements Serializable</h3>
                     <ul class="blockList">
                         <li class="blockList">
