+++ src/main/java/com/meti/client/ClientCreator.java
@@ -1,10 +1,9 @@
 package com.meti.client;
 
-import com.meti.util.Dialog;
-import com.meti.util.Utility;
+import com.meti.util.fx.Dialog;
+import com.meti.util.misc.Utility;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
@@ -15,8 +14,6 @@ import java.net.Socket;
 import java.net.URL;
 import java.util.concurrent.ExecutorService;
 import java.util.logging.Level;
-import java.util.logging.Logger;
-import java.util.logging.SimpleFormatter;
 
 import static com.meti.client.ClientProperties.logger;
