+++ src/main/java/com/meti/util/fx/FXBundle.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.fx;
 
 import javafx.scene.Parent;
 import javafx.stage.Stage;
