+++ src/main/java/com/meti/util/intf/Action.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.intf;
 
 import java.io.IOException;
