+++ src/main/java/com/meti/util/intf/Callback.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.intf;
 
 /**
  * @author SirMathhman
