+++ src/main/java/com/meti/util/io/Cargo.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.io;
 
 import java.io.Serializable;
 import java.util.ArrayList;
