+++ src/main/java/com/meti/util/io/Change.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.io;
 
 public class Change {
     //    TODO: handle change
