+++ src/main/resources/fxml/util/Dialog.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.*?>
 <AnchorPane xmlns:fx="http://javafx.com/fxml/1" prefHeight="400.0" prefWidth="600.0"
             xmlns="http://javafx.com/javafx/8"
-            fx:controller="com.meti.util.Dialog">
+            fx:controller="com.meti.util.fx.Dialog">
   <children>
     <ScrollPane fitToHeight="true" fitToWidth="true" prefViewportHeight="400.0"
       prefViewportWidth="600.0">
